Add files via upload
Both the data entered from the General Social Survey and the R code are listed here. This visualization consists of a line graph of the three different happiness levels recorded 
(Very Happy, Pretty Happy, and Not Too Happy) since roughly 1972 until 2021. The visualization also consists of an interactive bar chart to further compare the happiness levels for each respective year. The user is able to click on whatever recorded year they wish to have a separate bar chart constructed.